Stitch adjacent strings before transforming ellipsis
This allows properly replacing the double spaces with single spaces after transformation.
