+++ latex/filters/transform-ellipsis.lua
@@ -1,5 +1,5 @@
 function Str (el)
   if string.match(el.text, '…') then
-      return pandoc.Str(string.gsub(el.text, "…", ". . ."))
+      return pandoc.Str(string.gsub(string.gsub(el.text, "…", " . . . "), "  ", " "))
   end
 end
